Nothing got really done

+++ src/SchroedingerSolver/include/params/Params1D.hpp
@@ -6,9 +6,9 @@ class Params1D: Params{
 
 public:
 	Params1D(complex<double> xma,complex<double> xmi,complex<double> tma,complex<double> tmi,int gridX,int gridT, int gridE): xmax(xma),xmin(xmi),tmax(tma),tmin(tmi),nx(gridX),nt(gridT),ne(gridE){};
-	int getnx();
-	int getne();
-	int getnt();
+    size_t getnx();
+	size_t getne();
+	size_t getnt();
 	double getxmax();
 	double getxmin();
 	double gettmax();
@@ -18,7 +18,7 @@ complex<double> xmax;
 complex<double> xmin;
 complex<double> tmax;
 complex<double> tmin;
-int nx,nt,ne;
+size_t nx,nt,ne;
 
 };
 
@@ -26,7 +26,7 @@ double Params1D::getxmax(){return xmax.real();}
 double Params1D::getxmin(){return xmin.real();}
 double Params1D::gettmax(){return tmax.real();}
 double Params1D::gettmin(){return tmin.real();}
-int Params1D::getne(){return ne;}
-int Params1D::getnx(){return nx;}
-int Params1D::getnt(){return nt;}
+size_t Params1D::getne(){return ne;}
+size_t Params1D::getnx(){return nx;}
+size_t Params1D::getnt(){return nt;}
 #endif
